Adds PostgreSQL charm integration
Updates the metadata.yaml file to include a database relation.
Adds ops-lib-pgsql in requirements.txt, which we can use to handle
the PostgreSQL Charm relation.

The Waltz charm will prioritize the PostgreSQL database provisioned to us
by the charm. If there is no PostgreSQL charm relation, or if it was broken,
the database connection details present in the Waltz Charm configuration
will be used by Waltz instead.

Adds unit test to validate the scenario.